% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgets-package.R
\docType{package}
\name{gadgets}
\alias{gadgets}
\alias{gadgets-package}
\title{Gadgets}
\description{
Use RStudio addins to execute \R functions interactively within RStudio.
See \code{\link{googleSearchAddin}()}.
}
